Minor reorganization. Make behavior of $file input box more conistently controlled by the reply object. Make sure if it's visible, or actually in use for posting, then its value always matches with what's expected. If it's not visible and not in use, don't worry too much about clearing its old value. Maybe browsers will be more likely to browse from the same folder on its next use if it still has an old value in it.
Hide the Add reply button if queuing isn't in use.

Shift-click on file input box should remove the file, not the reply.

If the current reply's delete button is clicked, don't rely on the the
click event bubbling up to the preivew image and causing it to be
re-selected that way. In fact, explicitly disable click events on the
delete buttons from propagating so that way the event can never bubble
up to the preview image and attempt to select the just-deleted reply.