Temporary antispam fix.
Something about the antispam stuff is broken: for some reason,
legitimate clients are sending all of the hidden values correctly, but
with a hash value that matches up with a different set of hidden values.
NEED TO FIX

Until then, only reject bad antispam hashes if there's no userid. It
seems that none of the current spambots use javascript and none of them
send a userid. Nearly all legitimate users have javascript on.