Fix bug in watcher where if you viewed a post preview in the last post in a watched thread, then the watcher would later claim that there were an unknown number of new posts in the thread.
When you view a thread, the watcher remembers how many posts it saw in
the thread, and the time of the last post in the thread. When you viewed
a post link preview in the last post, the watcher would think that the
previewed post was the last post in the thread, and it would record that
post's time as the last post time seen. Later, the watcher queries the
server for the time of the latest post in the thread, sees that value as
being more recent, and reports that there are new posts in the thread.