Make sure js is forward compatible with jQuery v1.9 series.
State of checkboxes needs to be checked and modified with .prop(), not
.attr().

Mouse events don't have the .clientX and .clientY properties now, so the
QR dragging has been reworked slightly to make a bit more sense.

HTML needs to be parsed with $.parseHTML first, so that way it can't be
mistaken for a selector.