Respond to the QR with JSON instead of sending the thread immediately.
Previously, when someone made a post with the QR, it would send the post
request, which would then redirect the browser to the thread. The QR's
ajax call would see this as part of the same connection, and if that
second request to load the thread failed, the QR would report that there
was a connection error and respond as if the post didn't go through.

This change de-couples the posting and refreshing of the thread. The QR
can now quickly see when the post has been successful.