auto-reload.js: Never load an old post onto the page.
When on a view-last-50 page, if posts are deleted and the new
auto-loaded page has posts that weren't there before at the top, then
they used to get loaded into the current page at the bottom.

This change makes it so posts only get added to the page if their number
is higher than the last post's number.