Underp Tinyboard's post remembering feature.
When you submit a post, the client remembers the post in case it didn't
work, and only forgets the post when they server says it went through.
Tinyboard was bugged to only remember one post at a time, and to not
always show the remembered post, so this fixes that.