Fixup truncate().
Simplify needlessly complex broken tag removal regex.

Remove redundant check that the results from the tag detection regex are sane.

Add new tags to the front of $open_tags, so that way they're closed in
the correct order.

Make sure closed tags only remove the correct tag from the $open_tags
array. It used to be that if a tag was closed that hadn't been seen,
then the first tag in $open_tags would be unset.