Temporary work-around to prepare for removing "username" from config.spam.hidden_input_names.
When we remove "username" from config.spam.hidden_input_names, some
users will still have pages loaded with that used in some of the
anti-bot fields. Just disable anti-bot checking for people who submit
"username" as one of their anti-bot fields temporarily so we don't block
their posts. We can take this work-around off after a day after all the
pages are rebuilt to not have that in any of the anti-bot fields.