Fix issue where the previewed copy of a post could be used in another preview.
If you expanded a post's backlink, and then hovered over a link to that
backlinked post, then the hover preview would have a transparent
background because it would use the first found post in the page with
that id, which would be the one in the backlink preview. This commit
fixes it so toQuerySelector() returns a selector which would only match
the original post, not a preview.

Don't put the post id in the class of a dummy preview post. We don't
want a dummy preview being mistaken for the actual post. Not that this
is much of an issue since the above bug was fixed I guess.