Use Kefir streams in post preview code to simplify unregistering.
It's easier to see that there aren't any leaked event listeners this
way, and it avoids unregistering other possible event listeners on the
links.

onPostLinkEvent changed to return a stream so it can be unregistered.

post-inline and post-hover modules made to be live-updateable.

Fix issue I introduced with inline previews not being closeable.