New way of tracking database migrations.
This will fix the issue where multiple branches developed in parallel
that need their own database migrations would use the same version
numbers for their migrations and run into each other and be painful to
switch between or merge.

Now all migrations are named, and completed migrations are stored in the
database itself.