Add filter previewing system.
When the user makes a request to preview a filter, a task is started on
the server to do the search, and the server responds with the URL to the
task. The client then must poll the URL to see when the task is
complete. When it's complete, the task URL will redirect to the preview
results URL.

This means things won't break if the search takes longer than 30 seconds
to complete.

This follows the guidelines in
http://farazdagi.com/blog/2014/rest-long-running-jobs/